ZON-6478: Move existing teasers from AutomaticArea to Area
Moves the function existing_teasers to Area and adjusts some cache issues.
existing_teasers must be passed to AutomaticArea by its constructor so
that it is be accessible in ContentQueries.
TODO: Move parent_cache functionality to zeit.cms ?!
