+++ core/src/zeit/content/cp/area.py
@@ -1,6 +1,8 @@
 from zeit.cms.content.property import ObjectPathAttributeProperty
 from zeit.cms.i18n import MessageFactory as _
-from zeit.content.cp.interfaces import IAutomaticTeaserBlock, ITeaserBlock
+from zeit.content.cp.centerpage import writeabledict
+from zeit.content.cp.interfaces import (
+    IAutomaticTeaserBlock, ICenterPage, ITeaserBlock)
 import gocept.lxml.interfaces
 import grokcore.component as grok
 import lxml.etree
@@ -16,6 +18,27 @@ import zeit.edit.interfaces
 import zope.component
 import zope.container.interfaces
 import zope.interface
+
+
+def parent_cache(context, doc_iface, name, factory=writeabledict):
+    parent = doc_iface(context)
+    return parent.cache.setdefault(name, factory())
+
+
+def cached_on_parent(
+        doc_iface, attr=None, keyfunc=lambda x: x, factory=writeabledict):
+    """ Decorator to cache the results of the function in a dictionary
+        on the centerpage.  The dictionary keys are built using the optional
+        `keyfunc`, which is called with `self` as a single argument. """
+    def decorator(fn):
+        def wrapper(self, *args, **kw):
+            cache = parent_cache(self, doc_iface, attr or fn.__name__)
+            key = keyfunc(self)
+            if key not in cache:
+                cache[key] = fn(self, *args, **kw)
+            return cache[key]
+        return wrapper
+    return decorator
 
 
 @zope.component.adapter(
@@ -383,6 +406,38 @@ class Area(zeit.content.cp.blocks.block.VisibleMixin,
         for block in list(self.values()):
             if IAutomaticTeaserBlock.providedBy(block):
                 del self[block.__name__]
+
+    @property
+    @cached_on_parent(ICenterPage, keyfunc=lambda x: x.__name__)
+    def existing_teasers(self):
+        current_area = self
+        cp = ICenterPage(self)
+        area_teasered_content = parent_cache(
+            cp, ICenterPage, 'area_teasered_content')
+        area_manual_content = parent_cache(
+            cp, ICenterPage, 'area_manual_content')
+
+        seen = set()
+        above = True
+        for area in cp.cached_areas:
+            if area == current_area:
+                above = False
+            if above:  # automatic teasers above current area
+                if area not in area_teasered_content:
+                    area_teasered_content[area] = set(
+                        zeit.content.cp.interfaces.ITeaseredContent(area))
+
+                seen.update(area_teasered_content[area])
+            else:  # manual teasers below (or in) current area
+                if area not in area_manual_content:
+                    # Probably not worth a separate adapter (like
+                    # ITeaseredContent), since the use case is pretty
+                    # specialised.
+                    area_manual_content[area] = set(
+                        zeit.content.cp.blocks.teaser.extract_manual_teasers(
+                            area))
+                seen.update(area_manual_content[area])
+        return seen
 
     def filter_values(self, *interfaces):
         return zeit.content.cp.interfaces.IRenderedArea(self).filter_values(
