+++ core/src/zeit/content/cp/tests/test_automatic.py
@@ -525,7 +525,6 @@ class HideDupesTest(zeit.content.cp.testing.FunctionalTestCase):
 
     def create_automatic_area(self, cp, count=3, type='centerpage'):
         area = cp['feature'].create_item('area')
-        self.automatic_area = zeit.content.cp.automatic.AutomaticArea(area)
         area.count = count
         area.automatic_type = type
         area.automatic = True
@@ -734,13 +733,11 @@ class HideDupesTest(zeit.content.cp.testing.FunctionalTestCase):
 
     def test_teaser_count(self):
         a1 = self.create_automatic_area(self.cp, count=0, type='topicpage')
-        a1_automatic_area = self.automatic_area
         a2 = self.create_automatic_area(self.cp, count=0, type='topicpage')
         a3 = self.create_automatic_area(self.cp, count=2)
         a1.referenced_topicpage = 'tms-id'
         a3.referenced_topicpage = 'tms-id'
-        tms_query = zeit.contentquery.query.TMSContentQuery(
-            a1_automatic_area)
+        tms_query = zeit.contentquery.query.TMSContentQuery(a1)
         self.assertEqual(tms_query._teaser_count, 0)
         a2.count = 3
         self.assertEqual(tms_query._teaser_count, 0)
@@ -753,7 +750,6 @@ class HideDupesTest(zeit.content.cp.testing.FunctionalTestCase):
 
     def test_total_hits_can_be_called_first(self):
         area = self.create_automatic_area(self.cp)
-        automatic_area = self.automatic_area
         area.start = 0          # TODO: are we sure this is _always_ set?
         tms = mock.Mock()
         zope.component.getGlobalSiteManager().registerUtility(
@@ -761,8 +757,7 @@ class HideDupesTest(zeit.content.cp.testing.FunctionalTestCase):
         results = zeit.cms.interfaces.Result([])
         results.hits = 42
         tms.get_topicpage_documents.return_value = results
-        tms_query = zeit.contentquery.query.TMSContentQuery(
-            automatic_area)
+        tms_query = zeit.contentquery.query.TMSContentQuery(area)
         self.assertEqual(tms_query.total_hits, 42)
 
 
@@ -832,10 +827,8 @@ class AutomaticRSSTest(HideDupesTest):
 
     def test_rss_content_query_creates_teasers_from_feed(self):
         area = self.create_automatic_area(self.cp, count=3, type='rss-feed')
-        automatic_area = self.automatic_area
         m = self.mocked_rss_query(area)
-        rss_query = zeit.contentquery.query.RSSFeedContentQuery(
-            automatic_area)
+        rss_query = zeit.contentquery.query.RSSFeedContentQuery(area)
         with m:
             result = rss_query()
         self.assertEqual(3, len(result))
